update documentation remove TimeSeriesUtil.java and use TSUtil instead
